backend/vector-micro-service/pyproject.toml
[project]
name = "vector-micro-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.0.15",
    "fastapi>=0.116.1",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "pypdf>=5.7.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
